Extract lens drag-drop from lens editor app
