+++ x-pack/platform/plugins/shared/lens/test/scout/core/ui/parallel_tests/multiple_data_views.spec.ts
@@ -86,7 +86,7 @@ spaceTest.describe('Lens with multiple data views', { tag: '@local-stateful-clas
 
         // Flights layer — switch data panel first so the new layer inherits flights,
         // then add a line layer and toggle DistanceKilometers (matches FTR order).
-        await lens.switchDataPanelIndexPattern(testData.DATA_VIEW_ID.FLIGHTS);
+        await lens.dragDrop.switchDataPanelIndexPattern(testData.DATA_VIEW_ID.FLIGHTS);
         await addDataLayer(page, 'line');
         await lens.layers.activateLayerTab(1);
         await page.testSubj.locator('fieldToggle-DistanceKilometers').waitFor({ state: 'visible' });
